+++ lib/webspicy/checker.rb
@@ -8,18 +8,17 @@ module Webspicy
 
     def call
       config.each_scope do |scope|
-        client = scope.get_client
         scope.each_resource_file do |file, folder|
           RSpec.describe file.relative_to(folder).to_s do
 
             it 'meets the formal doc data schema' do
-              Webspicy.resource(file.load, file)
+              Webspicy.resource(file.load, file, scope)
             end
 
           end
         end
-        RSpec::Core::Runner.run config.rspec_options
       end
+      RSpec::Core::Runner.run config.rspec_options
     end
 
   end